Kill_Hitman/requirements.txt
keyboard
psutil
pystray
Pillow
pyinstaller
pywin32